Refactoring handler.ts to seperate handler files.
